database/scripts/requirements.txt
python-dateutil==2.9.0
pymongo==4.7.1